+++ packages/common/CHANGELOG.md
@@ -1,12 +1,17 @@
 # Changelog
 
 
+## 0.3.3 (2025-11-28)
+
+- Add support for Uniswap Hooks Wizard ([#628](https://github.com/OpenZeppelin/contracts-wizard/pull/628))
+
+## 0.3.2 (2025-11-26)
+
+- Add tokenUri setting for stellar non fungible model ([#725](https://github.com/OpenZeppelin/contracts-wizard/pull/725))
+
 ## 0.3.1 (2025-11-12)
 
 - Solidity account signer: Add `WebAuthn` to the list of signers available. ([#718](https://github.com/OpenZeppelin/contracts-wizard/pull/718))
-
-## 0.3.0 (2025-11-11)
-
 - **Breaking changes**: Solidity Stablecoin and RWA: Change `custodian` option to `freezable`. Replace ERC20Custodian with ERC20Freezable. ([#719](https://github.com/OpenZeppelin/contracts-wizard/pull/719))
 - Add macros descriptions for Cairo ([#703](https://github.com/OpenZeppelin/contracts-wizard/pull/703))