+++ packages/core/solidity/CHANGELOG.md
@@ -1,12 +1,17 @@
 # Changelog
 
 
+## 0.10.4 (2025-11-28)
+
+- Add support for Uniswap Hooks Wizard ([#628](https://github.com/OpenZeppelin/contracts-wizard/pull/628))
+
+## 0.10.3 (2025-11-26)
+
+- Fix missing dependency for get-versioned-remappings ([#740](https://github.com/OpenZeppelin/contracts-wizard/pull/740))
+
 ## 0.10.2 (2025-11-12)
 
 - Solidity account signer: Add `WebAuthn` to the list of signers available. ([#718](https://github.com/OpenZeppelin/contracts-wizard/pull/718))
-
-## 0.10.1 (2025-11-11)
-
 - Fixed bug with incorrect names in generated comment for Multisig account. ([#720](https://github.com/OpenZeppelin/contracts-wizard/pull/720))
 - Add API function to get versioned remappings. ([#724](https://github.com/OpenZeppelin/contracts-wizard/pull/724))
 - **Breaking changes**: Solidity Stablecoin and RWA: Change `custodian` option to `freezable`. Replace ERC20Custodian with ERC20Freezable. ([#719](https://github.com/OpenZeppelin/contracts-wizard/pull/719))